mshopa/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.0
django-allauth==0.51.0
django-filer==2.2.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-paypal==2.0
django-polymorphic==3.1.0
django-shopping-cart==0.1
easy-thumbnails==2.8.5
idna==3.4
lxml==4.9.2
oauthlib==3.2.1
Pillow==9.4.0
pip==22.3.1
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
setuptools==65.6.3
soupsieve==2.3.2.post1
sqlparse==0.4.2
svglib==1.5.1
tinycss2==1.2.1
types-cryptography==3.3.23
tzdata==2022.7
Unidecode==1.1.2
urllib3==1.26.12
webencodings==0.5.1
wheel==0.37.1
wincertstore==0.2
